add Q_PROPERTY to abstract_notification, small fixes

+++ plugins/logic/src/abstract_notification.cpp
@@ -52,7 +52,7 @@ namespace LogicPlugin {
         }
     }
 
-    void AbstractNotification::setId(int32_t id)
+    void AbstractNotification::setId(int16_t id)
     {
         _id = id;
     }
